A_lot_of_thinks_some_bugs_added_too

+++ Clowns' F8/Entity.cpp
@@ -92,7 +92,7 @@ bool Entity::CleanUp()
 	return true;
 }
 
-std::pair<float, float> Entity::GetPosition()
+std::pair<int, int> Entity::GetPosition()
 {
 	return position;
 }
